Добавление авторизации

+++ app/Http/Controllers/ReportController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\report;
 use App\Models\Status;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Auth;
 
 class ReportController extends Controller
 {
@@ -22,10 +22,12 @@ class ReportController extends Controller
         ]);
         if($validate){
             $reports = report::where('status_id', $status)
+                    ->where('user_id', Auth::user()->id)
                     ->orderBy('created_at', $sort)
                     ->simplePaginate(8);
         } else {
-            $reports = report::orderBy('created_at', $sort)
+            $reports = report::where('user_id', Auth::user()->id)
+                    ->orderBy('created_at', $sort)
                     ->simplePaginate(8);
         }
 
@@ -35,8 +37,13 @@ class ReportController extends Controller
     }
 
     public function destroy(Report $report){
-        $report->delete();
-        return redirect()->back();
+        if(Auth::user()->id === $report->user_id){
+            $report->delete();
+            return redirect()->back();
+        }
+        else{
+            abort(403, ' У вас нету прав для редактирования этой записи');
+        }
     }
 
     public function store(Request $request, report $report){
@@ -45,21 +52,34 @@ class ReportController extends Controller
             'description' => 'string',
         ]);
 
+        $data['user_id'] = Auth::user()->id;
+        $data['status_id'] = 1;
+
         $report->create($data);
         return redirect()->back();
     }
 
     public function edit(report $report){
-        return view('report.edit', compact('report'));
+        if(Auth::user()->id === $report->user_id){
+            return view('report.edit', compact('report'));
+        }
+        else{
+            abort(403, ' У вас нету прав для редактирования этой записи');
+        }
     }
 
     public function update(Request $request, report $report){
-        $data = $request -> validate([
-           'number' => 'string',
-           'description' => 'string', 
-        ]);
+        if(Auth::user()->id === $report->user_id){
+            $data = $request -> validate([
+            'number' => 'string',
+            'description' => 'string', 
+            ]);
 
-        $report ->update($data);
-        return redirect()->back();
+            $report ->update($data);
+            return redirect()->back();
+        }
+        else{
+            abort(403, ' У вас нету прав для редактирования этой записи');
+        }
     }
 }
